--- a/interpreter/interpreter_test.go
+++ b/interpreter/interpreter_test.go
@@ -883,8 +883,8 @@ func TestStack(t *testing.T) {
 			FunctionName: "window",
 			Location: ast.SourceLocation{
 				File:   "universe/universe.flux",
-				Start:  ast.Position{Line: 3768, Column: 12},
-				End:    ast.Position{Line: 3768, Column: 51},
+				Start:  ast.Position{Line: 3770, Column: 12},
+				End:    ast.Position{Line: 3770, Column: 51},
 				Source: `window(every: inf, timeColumn: timeDst)`,
 			},
 		},