--- a/libflux/go/libflux/buildinfo.gen.go
+++ b/libflux/go/libflux/buildinfo.gen.go
@@ -495,7 +495,7 @@ var sourceHashes = map[string]string{
 	"stdlib/universe/histogram_quantile_test.flux":                                                "8f6ef6a0c738bf70d77215857f3c3a677a1c2ff7c3dc818cfd3774d0fac32478",
 	"stdlib/universe/histogram_test.flux":                                                         "3016cb26398869dbaa2b7c0f35e2c99b105d3df411855da2d562eccefa3ff8a7",
 	"stdlib/universe/holt_winters_panic_test.flux":                                                "204eb8044d634e5350a364eac466eb51e7f549e4ac7f454de7b244ba272b248f",
-	"stdlib/universe/holt_winters_test.flux":                                                      "9f91192e791d4c523491ad243296da99043aedf88ac3c563d086731bc633d103",
+	"stdlib/universe/holt_winters_test.flux":                                                      "b7f017eb36fcfd52dea337a8394d26b799efa285e16f320bb569b8593c4d21fa",
 	"stdlib/universe/hour_selection_test.flux":                                                    "af919d0f515984e1155194a05dbe92d533c212bef1eb23e014e862ffb85810a5",
 	"stdlib/universe/increase_test.flux":                                                          "3e94fe06b5b71aadc5ac8b6f070278bba663a9633b4d4b3ec58903792bfe0bc8",
 	"stdlib/universe/integral_test.flux":                                                          "7206d881f059f0e6009fe3f0cfeb18b5fcec9ba192b607f63d375541851fcb85",