--- a/stdlib/universe/holt_winters/holt_winters.go
+++ b/stdlib/universe/holt_winters/holt_winters.go
@@ -7,6 +7,7 @@ import (
 	"github.com/influxdata/flux/array"
 	"github.com/influxdata/flux/arrow"
 	"github.com/influxdata/flux/internal/mutable"
+	"gonum.org/v1/gonum/optimize"
 )
 
 // HoltWinters forecasts a series into the future.
@@ -122,6 +123,18 @@ func (r *HoltWinters) Do(vs *array.Float) *array.Float {
 	// Determine best fit for the various parameters
 	minSSE := math.Inf(1)
 	var bestParams *mutable.Float64Array
+
+	// Params for gonum optimize
+	problem := optimize.Problem{
+		Func: func(par []float64) float64 {
+			f := mutable.NewFloat64Array(r.alloc)
+			defer f.Release()
+			f.AppendValues(par)
+			return r.sse(f)
+		},
+	}
+	settings := optimize.Settings{Converger: &optimize.FunctionConverge{Absolute: 1e-10, Iterations: 100}}
+
 	for alpha := hwGuessLower; alpha < hwGuessUpper; alpha += hwGuessStep {
 		for beta := hwGuessLower; beta < hwGuessUpper; beta += hwGuessStep {
 			for gamma := hwGuessLower; gamma < hwGuessUpper; gamma += hwGuessStep {
@@ -130,8 +143,18 @@ func (r *HoltWinters) Do(vs *array.Float) *array.Float {
 					initParams.Set(1, beta)
 					initParams.Set(2, gamma)
 					initParams.Set(3, phi)
-					// Optimize creates new parameters every time it is called.
-					sse, newParams := r.optim.Optimize(r.sse, initParams, r.epsilon, 1)
+
+					// Minimize creates new parameters every time it is called.
+					result, err := optimize.Minimize(problem, initParams.Float64Values(), &settings, &optimize.NelderMead{})
+					if err != nil {
+						panic(err)
+					}
+
+					f := mutable.NewFloat64Array(r.alloc)
+					defer f.Release()
+					f.AppendValues(result.X)
+					sse, newParams := result.F, f
+
 					if sse < minSSE || bestParams == nil {
 						if bestParams != nil {
 							// Previous bestParams are not the best anymore. We can release them.