Replaced our existing Nelder-Mead optimization with [gonum optimize](https://pkg.go.dev/gonum.org/v1/gonum@v0.11.0/optimize)

fixes: https://github.com/influxdata/flux/issues/4307

### Benchmark -
_go test -bench=BenchmarkHoltWinters_
goos: darwin
goarch: amd64
pkg: github.com/influxdata/flux/stdlib/universe
cpu: Intel(R) Core(TM) i7-9750H CPU @ 2.60GHz

### Existing [Nelder-mead](https://github.com/influxdata/flux/blob/master/stdlib/universe/holt_winters/neldermead.go) method implementation
```
BenchmarkHoltWintersWithoutFit-12                          34815             35703 ns/op            3608 B/op         60 allocs/op
BenchmarkHoltWintersWithFit-12                             37070             32289 ns/op            3608 B/op         60 allocs/op
BenchmarkHoltWintersWithoutFitSeasonality-12               38385             31545 ns/op            3608 B/op         60 allocs/op
BenchmarkHoltWintersWithFitSeasonality-12                  35692             31936 ns/op            3608 B/op         60 allocs/op
```

### Gonum [Nelder-mead](https://github.com/gonum/gonum/blob/master/optimize/neldermead.go) implementation
```
BenchmarkHoltWintersWithoutFit-12                          37761             30296 ns/op            3672 B/op         61 allocs/op
BenchmarkHoltWintersWithFit-12                             37507             30845 ns/op            3672 B/op         61 allocs/op
BenchmarkHoltWintersWithoutFitSeasonality-12               40724             30385 ns/op            3672 B/op         61 allocs/op
BenchmarkHoltWintersWithFitSeasonality-12                  36145             32799 ns/op            3672 B/op         61 allocs/op
```






### Done checklist
- [ ] docs/SPEC.md updated
- [x] Test cases written