@skartikey 2022-06-22
benchstat after this commit [perf: allocate fix size mem and copy values](https://github.com/influxdata/flux/pull/4816/commits/b596f90353540d94d13768ed616637cfdaa7c347)

_benchstat before.bench memFix.bench_
```
name                                 old time/op    new time/op    delta
HoltWintersWithoutFit-12               20.4ms ± 0%    17.6ms ± 0%   ~     (p=1.000 n=1+1)
HoltWintersWithFit-12                  19.6ms ± 0%    18.3ms ± 0%   ~     (p=1.000 n=1+1)
HoltWintersWithoutFitSeasonality-12    27.9µs ± 0%    31.7µs ± 0%   ~     (p=1.000 n=1+1)
HoltWintersWithFitSeasonality-12       28.7µs ± 0%    31.5µs ± 0%   ~     (p=1.000 n=1+1)

name                                 old alloc/op   new alloc/op   delta
HoltWintersWithoutFit-12               4.92MB ± 0%    2.46MB ± 0%   ~     (p=1.000 n=1+1)
HoltWintersWithFit-12                  4.92MB ± 0%    2.46MB ± 0%   ~     (p=1.000 n=1+1)
HoltWintersWithoutFitSeasonality-12    3.67kB ± 0%    3.67kB ± 0%   ~     (all equal)
HoltWintersWithFitSeasonality-12       3.67kB ± 0%    3.67kB ± 0%   ~     (all equal)

name                                 old allocs/op  new allocs/op  delta
HoltWintersWithoutFit-12                57.8k ± 0%     28.9k ± 0%   ~     (p=1.000 n=1+1)
HoltWintersWithFit-12                   57.8k ± 0%     28.9k ± 0%   ~     (p=1.000 n=1+1)
HoltWintersWithoutFitSeasonality-12      61.0 ± 0%      61.0 ± 0%   ~     (all equal)
HoltWintersWithFitSeasonality-12         61.0 ± 0%      61.0 ± 0%   ~     (all equal)
```