@skartikey 2022-06-22
Ran profiler on benchmark - 

**CPU -** 

![image](https://user-images.githubusercontent.com/1942366/175059705-c57e77a0-4805-4056-8f84-ebe9f1ae9b27.png)

![pprof-cpu](https://user-images.githubusercontent.com/1942366/175062118-1b21e526-4f66-4ef8-8f64-45e9910a539a.svg)


**Mem -**
![image](https://user-images.githubusercontent.com/1942366/175059948-a986bca7-9c28-4569-966f-b5824bcceb4a.png)
![pprof-mem](https://user-images.githubusercontent.com/1942366/175062358-082e681e-91e8-40d2-bab1-af0465d3a7dc.svg)