@skartikey 2022-06-23
benchstat after this commit [perf: more mem allocation fix](https://github.com/influxdata/flux/pull/4816/commits/c4224b4b20b0e5b898546aced3969e4241859000)

_benchstat memFix.bench moreMemFix.bench_
```
name                                 old time/op    new time/op    delta
HoltWintersWithoutFit-12               17.6ms ± 0%     0.0ms ± 0%   ~     (p=1.000 n=1+1)
HoltWintersWithFit-12                  18.3ms ± 0%     0.0ms ± 0%   ~     (p=1.000 n=1+1)
HoltWintersWithoutFitSeasonality-12    31.7µs ± 0%    28.8µs ± 0%   ~     (p=1.000 n=1+1)
HoltWintersWithFitSeasonality-12       31.5µs ± 0%    30.9µs ± 0%   ~     (p=1.000 n=1+1)

name                                 old alloc/op   new alloc/op   delta
HoltWintersWithoutFit-12               2.46MB ± 0%    0.00MB ± 0%   ~     (p=1.000 n=1+1)
HoltWintersWithFit-12                  2.46MB ± 0%    0.00MB ± 0%   ~     (p=1.000 n=1+1)
HoltWintersWithoutFitSeasonality-12    3.67kB ± 0%    3.67kB ± 0%   ~     (all equal)
HoltWintersWithFitSeasonality-12       3.67kB ± 0%    3.67kB ± 0%   ~     (all equal)

name                                 old allocs/op  new allocs/op  delta
HoltWintersWithoutFit-12                28.9k ± 0%      0.1k ± 0%   ~     (p=1.000 n=1+1)
HoltWintersWithFit-12                   28.9k ± 0%      0.1k ± 0%   ~     (p=1.000 n=1+1)
HoltWintersWithoutFitSeasonality-12      61.0 ± 0%      61.0 ± 0%   ~     (all equal)
HoltWintersWithFitSeasonality-12         61.0 ± 0%      61.0 ± 0%   ~     (all equal)
```