@skartikey 2022-06-24
@nathanielc Here is the `minSSE` value I am getting running the [go test](https://github.com/influxdata/flux/blob/master/stdlib/universe/holt_winters_test.go#L166) with our existing and `gonum` implementation. Looking at the data below, the new one is returning better results for the `minSSE`. 

**Before -** 
minSSE:  20.208541132811906
minSSE:  20.208541132811906
minSSE:  20.208541132811906
minSSE:  20.208541132811906
minSSE:  20.208541132811906
minSSE:  20.208541132811906
minSSE:  20.208541132811906
minSSE:  38.22969757997137
minSSE:  55.02308769115148
minSSE:  30.561795293721094

**After -** 
minSSE:  3.963712566228843
minSSE:  3.963712566228843
minSSE:  3.963712566228843
minSSE:  3.963712566228843
minSSE:  3.963712566228843
minSSE:  3.963712566228843
minSSE:  3.963712566228843
minSSE:  55.00975192379251
minSSE:  30.52018686151099

 I am using a print statement to get these values.