@nathanielc 2022-06-24
@skartikey The SSE's look good. Is there not an easy way to include them in the Go tests?

For the most recent benchmark results it looks like some of the tests run in 0s. That seems suspicious any ideas why those benchmarks are reporting 0s?