@skartikey 2022-06-27
> @skartikey The SSE's look good. Is there not an easy way to include them in the Go tests?

to pass the SSE in the go test, I have to make SSE part of the public API result. Not sure if that would be a good idea.

> For the most recent benchmark results it looks like some of the tests run in 0s. That seems suspicious any ideas why 
those benchmarks are reporting 0s?

Looks like `benchstat` tool is doing something here. I ran benchmark again and it looks alright to me. 

**Before -** 

```
BenchmarkHoltWintersWithoutFit-12                          29800             39355 ns/op            3608 B/op         60 allocs/op
BenchmarkHoltWintersWithFit-12                             28484             36553 ns/op            3608 B/op         60 allocs/op
BenchmarkHoltWintersWithoutFitSeasonality-12               31322             40150 ns/op            3608 B/op         60 allocs/op
BenchmarkHoltWintersWithFitSeasonality-12                  28293             38462 ns/op            3608 B/op         60 allocs/op
```

**After -**

```
BenchmarkHoltWintersWithoutFit-12                          37884             55364 ns/op            3672 B/op         61 allocs/op
BenchmarkHoltWintersWithFit-12                             33810             33034 ns/op            3672 B/op         61 allocs/op
BenchmarkHoltWintersWithoutFitSeasonality-12               35882             34245 ns/op            3672 B/op         61 allocs/op
BenchmarkHoltWintersWithFitSeasonality-12                  30438             36250 ns/op            3672 B/op         61 allocs/op
```