@skartikey 2022-07-05
@nathanielc 
Debugged the [test case with nulls and missing data](https://github.com/influxdata/flux/blob/master/stdlib/universe/holt_winters_test.go#L680) today

**Input to holt winter algorithm from the test case -** 
```
seasonal 4
[]float64 [4.948,2.192,0,2.93,5.121,0,0,2.877,5.449,0.896,0,0,0,0.404,4.357,2.618,6.102,0,4.816]
```

What's happening is, in the setup phase of the algo, as [vs.Value(2)](https://github.com/influxdata/flux/blob/master/stdlib/universe/holt_winters/holt_winters.go#L114) is invalid according to the [rule](https://github.com/influxdata/flux/blob/master/stdlib/universe/holt_winters.go#L254) (Rows that have a null value are considered invalid, but used by the algorithm), initParams[8] is getting set to 0, 
Due to that, `stm` is getting set to 0 [here](https://github.com/influxdata/flux/blob/master/stdlib/universe/holt_winters/holt_winters.go#L222), which is setting[ lT to +Inf](https://github.com/influxdata/flux/blob/master/stdlib/universe/holt_winters/holt_winters.go#L164) and that makes bT to `Nan (Inf - Inf)`
This is introducing `Nan` in the algorithm which is messing up the rest of the processing. 
we need to discuss a way to handle the invalid input.